% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Les trajets par semaine}
\usage{
calcul_distribution_semaine(trajet, filtre)
}
\arguments{
\item{trajet}{A data.frame}
}
\value{
A data.frame
}
\description{
Les trajets par semaine
}
\examples{
 data(df_small)
calcul_distribution_semaine(df_small)

}
