% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_boucle}
\alias{compter_nombre_boucle}
\title{Compte le nombre de boucle}
\usage{
compter_nombre_boucle(trajet)
}
\arguments{
\item{trajet}{A data.frame contenant une colonne 'Numéro de boucle'}
}
\value{
Un entier correspondant au nombre de boucles distinctes
}
\description{
Compte le nombre de boucle
}
\examples{
data(df_small)
compter_nombre_boucle(df_small)
}
