% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compte le nombre de trajet}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{A data.frame}
}
\value{
A data.frame
}
\description{
Compte le nombre de trajet
}
\examples{
 data(df_small)
compter_nombre_trajets(df_small)

}
