% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\docType{data}
\name{df_small}
\alias{df_small}
\title{Petit jeu de données de trajets vélo}
\format{
A data.frame with variables:
\describe{
\item{Numéro de boucle}{Identifiant de la boucle}
\item{Jour de la semaine}{Jour de circulation}
\item{Total}{Nombre total de trajets}
\item{Probabilité de présence d'anomalies}{Indicateur d'anomalie}
}
}
\usage{
df_small
}
\description{
Un échantillon de données utilisé pour illustrer les fonctions du package.
}
\keyword{datasets}
