% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Jeu de données de trajets vélo sur les vacances de Toussaint}
\format{
A data.frame with variables:
\describe{
\item{Numéro de boucle}{Identifiant de la boucle}
\item{Jour de la semaine}{Jour de circulation}
\item{Total}{Nombre total de trajets}
\item{Probabilité de présence d'anomalies}{Indicateur d'anomalie}
}
}
\usage{
df_velo
}
\description{
La base dont on aura besoin dans le package.
}
\keyword{datasets}
