% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtre les anomalies non nulles}
\usage{
filtre_anomalie(trajet)
}
\arguments{
\item{trajet}{A data.frame}
}
\value{
A data.frame
}
\description{
Filtre les anomalies non nulles
}
\examples{
data(df_small)
filtre_anomalie(df_small)

}
