% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets par numéro de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets.}

\item{boucle}{Un vecteur de numéros de boucle à conserver.}
}
\value{
Un data.frame filtré sur les boucles sélectionnées.
}
\description{
Cette fonction permet de sélectionner uniquement les lignes
correspondant à une ou plusieurs boucles dans un jeu de données.
}
\examples{
filtrer_trajet(trajet = df_velo, boucle = c("880", "881"))
}
