% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Graphique}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{A data.frame}
}
\value{
A data.frame
}
\description{
Graphique
}
\examples{
data(df_small)
plot_distribution_semaine(df_small)

}
