% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le jour de plus de trajets}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{A data.frame}
}
\value{
A data.frame
}
\description{
Trouver le jour de plus de trajets
}
\examples{
data(df_small)
trouver_trajet_max(df_small)
}
